Adjust solution and test project organization - Organize solution by dividing projects into solution folders - Split the native testing across several files (partial class)
