+++ src/RSML/Evaluation/Evaluator.cs
@@ -35,7 +35,6 @@
  */
 
 using System;
-using System.Collections.Generic;
 
 using OceanApocalypseStudios.RSML.Analyzer;
 using OceanApocalypseStudios.RSML.Analyzer.Semantics;
