+++ tests/GlobalSuppressions.cs
@@ -3,5 +3,3 @@
 // Project-level suppressions either have no target or are given
 // a specific target and scoped to a namespace, type, member, etc.
 
-using System.Diagnostics.CodeAnalysis;
-
